Refactor integrations.neon
